+++ tests/at_end2end_test/test/pq/legacy_server_abort_test.dart
@@ -67,7 +67,8 @@ void main() {
             appName: 'b01-$label',
             deviceName: 'b01-$label-$runId',
             namespaces: namespaces,
-            otp: otp),
+            otp: otp,
+            signingAlgo: SigningAlgoType.rsa2048),
         AtLookupImpl(atSign, ConfigUtil.getYaml()['root_server']['url'],
             ConfigUtil.getYaml()['root_server']['port'] ?? 64));
     final record = (await owner.enrollmentService!.fetchEnrollmentRequests())